+++ src/annbatch/samplers/__init__.py
@@ -1,6 +1,11 @@
-from ._chunk_sampler import ChunkSampler, ChunkSamplerDistributed
+from ._chunk_sampler import ChunkSampler
+from ._distributed_sampler import DistributedSampler
+from ._random_sampler import RandomSampler
+from ._sequential_sampler import SequentialSampler
 
 __all__ = [
     "ChunkSampler",
-    "ChunkSamplerDistributed",
+    "DistributedSampler",
+    "RandomSampler",
+    "SequentialSampler",
 ]